refactor: CardFormContext 개선
- util, reducer, context 코드 분리
- context value의 non-nullable을 보장하는 createProvidedContext 구현
